Firehose: Refactor recordcache (#4176)
* Refactor recordcache for reuse in firehose client

* Reexport recordcache types in kinesis client package
