--- AmplifyClients/AmplifyKinesisClient/Sources/Support/SQLiteRecordStorage.swift
@@ -1,273 +1,0 @@
-//
-// Copyright Amazon.com Inc. or its affiliates.
-// All Rights Reserved.
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-import Foundation
-@preconcurrency import SQLite
-
-/// Uses actor for thread-safe access (Swift's recommended approach)
-actor SQLiteRecordStorage: RecordStorage {
-    private let database: Connection
-    private let cacheMaxBytes: Int64
-    private let maxRecords: Int
-    private let maxRecordSizeBytes: Int64
-    private let maxBytesPerStream: Int64
-    private let maxPartitionKeyLength: Int
-    private let identifier: String
-    private var cachedSize: Int64 = 0
-
-    // Table and columns - defined as static to avoid duplication in setupDatabase
-    private static let records = Table("records")
-    private static let id = Expression<Int64>("id")
-    private static let streamName = Expression<String>("stream_name")
-    private static let partitionKey = Expression<String>("partition_key")
-    private static let data = Expression<Data>("data")
-    private static let dataSize = Expression<Int>("data_size")
-    private static let retryCount = Expression<Int>("retry_count")
-    private static let createdAt = Expression<Double>("created_at")
-
-    init(
-        identifier: String,
-        maxRecords: Int,
-        cacheMaxBytes: Int64,
-        maxRecordSizeBytes: Int64,
-        maxBytesPerStream: Int64,
-        maxPartitionKeyLength: Int,
-        connection: Connection? = nil
-    ) throws {
-        self.identifier = identifier
-        self.maxRecords = maxRecords
-        self.cacheMaxBytes = cacheMaxBytes
-        self.maxRecordSizeBytes = maxRecordSizeBytes
-        self.maxBytesPerStream = maxBytesPerStream
-        self.maxPartitionKeyLength = maxPartitionKeyLength
-
-        let db = try connection ?? Self.createFileConnection(identifier: identifier)
-        self.database = db
-
-        try Self.setupSchema(on: db)
-        let size = try Self.wrapDatabaseError {
-            try db.scalar(Self.records.select(Self.dataSize.sum)) ?? 0
-        }
-        self.cachedSize = Int64(size)
-    }
-
-    private static func createFileConnection(identifier: String) throws -> Connection {
-        guard let path = try? FileManager.default.url(
-            for: .applicationSupportDirectory,
-            in: .userDomainMask,
-            appropriateFor: nil,
-            create: true
-        ) else {
-            throw RecordCacheError.database(
-                "Failed to locate application support directory",
-                defaultRecoverySuggestion
-            )
-        }
-
-        let dbPath = path.appendingPathComponent("kinesis_records_\(identifier).db").path
-        return try wrapDatabaseError {
-            try Connection(dbPath)
-        }
-    }
-
-    /// Sets up the database schema (tables and indices)
-    private static func setupSchema(on connection: Connection) throws {
-        try wrapDatabaseError {
-            try connection.run(records.create(ifNotExists: true) { table in
-                table.column(id, primaryKey: .autoincrement)
-                table.column(streamName)
-                table.column(partitionKey)
-                table.column(data)
-                table.column(dataSize)
-                table.column(retryCount, defaultValue: 0)
-                table.column(createdAt, defaultValue: Date().timeIntervalSince1970)
-            })
-
-            try connection.execute("CREATE INDEX IF NOT EXISTS idx_stream_id ON records (stream_name, id)")
-            try connection.execute("CREATE INDEX IF NOT EXISTS idx_data_size ON records (data_size)")
-        }
-    }
-
-    func addRecord(_ input: RecordInput) throws {
-        // Validate partition key length (1–256 Unicode scalars)
-        let partitionKeyScalarCount = input.partitionKey.unicodeScalars.count
-        if partitionKeyScalarCount == 0 || partitionKeyScalarCount > maxPartitionKeyLength {
-            throw RecordCacheError.validation(
-                "Partition key length \(partitionKeyScalarCount) is outside the allowed range of 1–\(maxPartitionKeyLength) characters",
-                "Use a partition key between 1 and \(maxPartitionKeyLength) characters."
-            )
-        }
-
-        // Validate per-record size limit (partition key + data blob)
-        if Int64(input.dataSize) > maxRecordSizeBytes {
-            throw RecordCacheError.validation(
-                "Record size \(input.dataSize) bytes exceeds the maximum of \(maxRecordSizeBytes) bytes (partition key + data blob)",
-                "Reduce the size of the data blob or partition key so their combined size does not exceed \(maxRecordSizeBytes) bytes."
-            )
-        }
-
-        // Check cache size limit before adding
-        if cachedSize + Int64(input.dataSize) > cacheMaxBytes {
-            throw RecordCacheError.limitExceeded(
-                "Cache size limit exceeded: \(cachedSize + Int64(input.dataSize)) bytes > \(cacheMaxBytes) bytes",
-                "Call flush() to send cached records or increase the cacheMaxBytes option."
-            )
-        }
-
-        let insert = Self.records.insert(
-            Self.streamName <- input.streamName,
-            Self.partitionKey <- input.partitionKey,
-            Self.data <- input.data,
-            Self.dataSize <- input.dataSize,
-            Self.retryCount <- 0,
-            Self.createdAt <- Date().timeIntervalSince1970
-        )
-
-        _ = try Self.wrapDatabaseError {
-            try database.run(insert)
-        }
-        cachedSize += Int64(input.dataSize)
-    }
-
-    func getRecordsByStream(afterIdByStream: [String: Int64] = [:]) throws -> [[Record]] {
-        try Self.wrapDatabaseError {
-            // Build per-stream WHERE clauses: id > lastProcessedId for streams we've already seen
-            let streamFilter: String
-            if afterIdByStream.isEmpty {
-                streamFilter = ""
-            } else {
-                let conditions = afterIdByStream.map { _ in
-                    "NOT (stream_name = ? AND id <= ?)"
-                }.joined(separator: " AND ")
-                streamFilter = "WHERE \(conditions)"
-            }
-
-            // Must use raw SQL for window functions
-            let query = """
-                SELECT id, stream_name, partition_key, data, data_size, retry_count, created_at
-                FROM (
-                    SELECT *,
-                           ROW_NUMBER() OVER (PARTITION BY stream_name ORDER BY id) as rn,
-                           SUM(data_size) OVER (PARTITION BY stream_name ORDER BY id) as running_size
-                    FROM records
-                    \(streamFilter)
-                )
-                WHERE rn <= ? AND running_size <= ?
-                ORDER BY stream_name, id
-                """
-
-            // Bind per-stream after-id filters
-            var bindings: [Binding?] = afterIdByStream.flatMap { (streamName, afterId) in
-                [streamName as Binding?, afterId as Binding?]
-            }
-            bindings.append(maxRecords as Binding?)
-            bindings.append(maxBytesPerStream as Binding?)
-
-            var recordsByStream: [String: [Record]] = [:]
-
-            for row: Statement.Element in try database.prepare(query, bindings) {
-                // Ensure row has expected number of columns (in case DB format is invalid)
-                guard row.count >= 7 else {
-                    throw RecordCacheError.database(
-                        "Unexpected row format: expected 7 columns, got \(row.count)",
-                        defaultRecoverySuggestion
-                    )
-                }
-
-                guard let id = row[0] as? Int64,
-                      let streamName = row[1] as? String,
-                      let partitionKey = row[2] as? String,
-                      let blob = row[3] as? SQLite.Blob,
-                      let retryCount = row[5] as? Int64,
-                      let createdAt = row[6] as? Double else {
-                    throw RecordCacheError.database(
-                        "Failed to parse record from database",
-                        defaultRecoverySuggestion
-                    )
-                }
-
-                let record = Record(
-                    id: id,
-                    streamName: streamName,
-                    partitionKey: partitionKey,
-                    data: Data(blob.bytes),
-                    retryCount: Int(retryCount),
-                    createdAt: Date(timeIntervalSince1970: createdAt)
-                )
-
-                recordsByStream[record.streamName, default: []].append(record)
-            }
-
-            return Array(recordsByStream.values)
-        }
-    }
-
-    func deleteRecords(ids: [Int64]) throws {
-        guard !ids.isEmpty else { return }
-
-        try Self.wrapDatabaseError {
-            let placeholders = ids.map { _ in "?" }.joined(separator: ",")
-            let sql = "DELETE FROM records WHERE id IN (\(placeholders))"
-
-            let statement = try database.prepare(sql)
-            try statement.run(ids.map { $0 as Binding })
-        }
-
-        try resetCacheSizeFromDb()
-    }
-
-    func incrementRetryCount(ids: [Int64]) throws {
-        guard !ids.isEmpty else { return }
-
-        try Self.wrapDatabaseError {
-            let placeholders = ids.map { _ in "?" }.joined(separator: ",")
-            let sql = "UPDATE records SET retry_count = retry_count + 1 WHERE id IN (\(placeholders))"
-
-            let statement = try database.prepare(sql)
-            try statement.run(ids.map { $0 as Binding })
-        }
-    }
-
-    func clearRecords() throws -> Int {
-        let count = try Self.wrapDatabaseError {
-            let count = try database.scalar(Self.records.count)
-            try database.run(Self.records.delete())
-            return count
-        }
-        cachedSize = 0
-        return count
-    }
-
-    func getCurrentCacheSize() throws -> Int64 {
-        return cachedSize
-    }
-
-    /// Resets the cached size by recalculating from the database
-    /// Used internally after delete operations to ensure accuracy
-    private func resetCacheSizeFromDb() throws {
-        let size = try Self.wrapDatabaseError {
-            try database.scalar(Self.records.select(Self.dataSize.sum)) ?? 0
-        }
-        cachedSize = Int64(size)
-    }
-
-    /// Wraps a throwing closure so that any non-RecordCacheError is converted
-    /// to `.database` with the underlying SQLite error attached.
-    private static func wrapDatabaseError<T>(_ operation: () throws -> T) throws -> T {
-        do {
-            return try operation()
-        } catch let error as RecordCacheError {
-            throw error
-        } catch {
-            throw RecordCacheError.database(
-                "A database error occurred",
-                defaultRecoverySuggestion,
-                error
-            )
-        }
-    }
-}
