+++ AmplifyClients/AmplifyRecordCache/Sources/RecordCacheError.swift
@@ -8,12 +8,9 @@
 import AmplifyFoundation
 import Foundation
 
-/// Default recovery suggestion for errors.
-let defaultRecoverySuggestion: RecoverySuggestion = "Inspect the underlying error for more details."
-
 /// Internal error type used by RecordClient/RecordStorage.
-/// Mapped to KinesisError via ``KinesisError/from(_:)``.
-enum RecordCacheError: Error {
+/// Mapped to a client-specific error type (e.g. KinesisError, FirehoseError).
+public enum RecordCacheError: Error {
     /// Database operation failed.
     case database(ErrorDescription, RecoverySuggestion, Error? = nil)
     /// Cache limit exceeded — no space for new records.
