+++ AmplifyClients/AmplifyRecordCache/Sources/RecordStorage.swift
@@ -5,7 +5,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-protocol RecordStorage: Actor {
+public protocol RecordStorage: Actor {
     /// Adds a new record to storage
     func addRecord(_ input: RecordInput) async throws
